+++ app/gui/src/project-view/components/RecordControl.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import SvgButton from '@/components/SvgButton.vue'
-import { useProjectStore } from '@/stores/project'
+import { useNodeExecution } from '@/stores/project/nodeExecution'
 import ControlButtons from './ControlButtons.vue'
 
-const project = useProjectStore()
+const nodeExecution = useNodeExecution()
 </script>
 
 <template>
@@ -14,7 +14,7 @@ const project = useProjectStore()
         class="iconButton"
         name="refresh"
         draggable="false"
-        @click.stop="project.executionContext.recompute()"
+        @click.stop="nodeExecution.recomputeAll()"
       />
     </template>
     <template #right>
@@ -23,7 +23,7 @@ const project = useProjectStore()
         class="iconButton"
         name="workflow_play"
         draggable="false"
-        @click.stop="project.executionContext.recompute('all', 'Live')"
+        @click.stop="nodeExecution.recomputeAll('Live')"
       />
     </template>
   </ControlButtons>