+++ app/ide-desktop/client/tests/createNewProject.spec.ts
@@ -3,7 +3,7 @@
 import { expect } from '@playwright/test'
 import { electronTest, loginAsTestUser } from './electronTest'
 
-electronTest('Create new project', async page => {
+electronTest('Create new project', async ({ page }) => {
   await loginAsTestUser(page)
   await expect(page.getByRole('button', { name: 'New Project', exact: true })).toBeVisible()
   await page.getByRole('button', { name: 'New Project', exact: true }).click()