+++ app/ide-desktop/client/tests/electronTest.ts
@@ -1,6 +1,9 @@
 /** @file Commonly used functions for electron tests */
 
 import { _electron, expect, type Page, test } from '@playwright/test'
+import fs from 'node:fs/promises'
+import os from 'node:os'
+import pathModule from 'node:path'
 
 const LOADING_TIMEOUT = 10000
 
@@ -9,7 +12,10 @@ const LOADING_TIMEOUT = 10000
  *
  * Similar to playwright's test, but launches electron, and passes Page of the main window.
  */
-export function electronTest(name: string, body: (page: Page) => Promise<void> | void) {
+export function electronTest(
+  name: string,
+  body: (args: { page: Page; projectsDir: string }) => Promise<void> | void,
+) {
   test(name, async () => {
     const app = await _electron.launch({
       executablePath: process.env.ENSO_TEST_EXEC_PATH ?? '',
@@ -20,7 +26,8 @@ export function electronTest(name: string, body: (page: Page) => Promise<void> |
     // Wait until page will be finally loaded: we expect login screen.
     // There's bigger timeout, because the page may load longer on CI machines.
     await expect(page.getByText('Login to your account')).toBeVisible({ timeout: LOADING_TIMEOUT })
-    await body(page)
+    const projectsDir = pathModule.join(os.tmpdir(), 'enso-test-projects', name)
+    await body({ page, projectsDir })
     await app.close()
   })
 }
@@ -51,4 +58,22 @@ export async function loginAsTestUser(page: Page) {
     await checkbox.click()
   }
   await page.getByRole('button').click()
+}
+
+/**
+ * Find the most recently edited Enso project in `projectsDir` and return its absolute path.
+ * There can be multiple projects, as the directory can be reused by subsequent test runs.
+ * We precisely know the naming schema for new projects, and we use this knowledge to
+ * find the project that was created most recently.
+ */
+export async function findMostRecentlyCreatedProject(projectsDir: string): Promise<string | null> {
+  const dirContent = await fs.readdir(projectsDir)
+  const sorted = dirContent.sort((a, b) => {
+    // Project names always end with a number, so we can sort them by that number.
+    const numA = parseInt(a.match(/\d+/)![0], 10)
+    const numB = parseInt(b.match(/\d+/)![0], 10)
+    return numA - numB
+  })
+  const last = sorted.pop()
+  return last != null ? pathModule.join(projectsDir, last) : null
 }